♻️ refactor: 공고 카드 컴포넌트 분리
