+++ app/(route)/(alba)/announce-list/page.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import CustomizedAnnounceCard from "./_components/customized-announce-card";
+import CustomizedAnnounce from "./_components/customized-announce";
 import { getAllNotices, getCustomizedNotices } from "@/app/_apis/api";
-import AllAnnounceCard from "./_components/all-announce-card";
+import AnnounceCard from "@/app/_components/announce-card";
 // import Filter from "./_components/filter";
 
 export default async function page() {
@@ -15,7 +15,7 @@ export default async function page() {
           <h2 className="pb-4 font-bold text-l md:pb-12 md:text-2xl">
             맞춤 공고
           </h2>
-          <CustomizedAnnounceCard notices={customizedNotices} />
+          <CustomizedAnnounce notices={customizedNotices} />
         </div>
       </div>
       <div className="mx-auto flex w-full flex-col px-2 md:justify-center lg:max-w-[964px]">
@@ -23,7 +23,9 @@ export default async function page() {
           전체 공고
         </h2>
         {/* <Filter></Filter> */}
-        <AllAnnounceCard notices={allNotices} />
+        <div className="lg grid grid-cols-2 gap-4 lg:grid-cols-3">
+          <AnnounceCard notices={allNotices} />
+        </div>
       </div>
     </div>
   );
